feat: add @pgpmjs/slice — plan slicing moved out of core, with opt-in AST dependency closure
