+++ pgpm/core/src/index.ts
@@ -1,5 +1,4 @@
 export * from './core/class/pgpm';
-export * from './slice';
 export * from './rebundle';
 export * from './extensions/extensions';
 export * from './modules/modules';
