+++ pgpm/slice/src/closure.ts
