+++ pgpm/slice/src/index.ts
