+++ pgpm/slice/src/output.ts
