+++ pgpm/slice/src/slice.ts
